requirements.txt
Django==4.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
environs==10.3.0
urllib3==1.26.6
celery==5.3.6
redis==5.0.1